+++ .vscode/enverant.json
@@ -1,11 +1,11 @@
 {
     # Define path to Freelancer Folder root
-    # "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/apps/freelancer_related/wine_prefix_freelancer_online/drive_c/Discovery",
+    "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/apps/freelancer_related/wine_prefix_freelancer_online/drive_c/Discovery",
     # "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/apps/freelancer_related/winprefix_freelancer_hd/drive_c/Freelancer HD Edition",
     # "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/windows10shared/game-repository",
     # "DARKSTAT_TRADE_ROUTES_DETAILED_TRADE_LANE": "true",
     # "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/windows10shared/fl-files-vanilla",
-    "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/apps/freelancer_related/wineprefix_flsr/drive_c/Freelancer",
+    # "CONFIGS_FREELANCER_FOLDER": "${env:HOME}/apps/freelancer_related/wineprefix_flsr/drive_c/Freelancer",
     "UTILS_DEV_ENV": "true",
     # "UTILS_BAD_VAR": "true",
     "DARKSTAT_LOG_LEVEL": "WARN",
